windsurf-project/backend/requirements.txt
# Core FastAPI and async framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
aiomysql==0.2.0

# Data validation and settings (compatible with opentrons)
# IMPORTANT: Pydantic must be kept below 2.0.0 for compatibility with opentrons library
# The opentrons package has not yet been updated to support Pydantic 2.x
pydantic>=1.10.9,<2.0.0

# WebSocket and HTTP clients
websockets==12.0
aiohttp==3.9.1
httpx==0.25.2

# MCP client dependencies
websockets==12.0  # Already included above for MCP WebSocket communication

# Configuration and environment
python-dotenv==1.0.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Logging
loguru==0.7.2

# Robot control libraries
mecademicpy>=1.4.0
opentrons==7.2.2
pyserial==3.5

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Remove asyncio package - use Python built-in
# asyncio==3.4.3  # REMOVED - conflicts with Python 3.7+ built-in